backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
langchain
openai
